Refactor test name
